fix Mandragora -> Dragon Toadstool
I'm used to the modern localization where マンドラゴラ ("Mandragora") in Japanese is referred to in English as Mandragora. However, since this project is using older localization standards, that being those seen in 1st gen MH (MH1, MHF1), the term "Dragon Toadstool" is used instead. This caused a mismatch between the item name and the item being referred to in dialog.