fix the smithy for the 15 trillionth time :')
lines 18-25 are gonna be the death of me, I swear...
for anyone reading here, this is what the issue is:
line 18 gets spoken before lines 19, 20, 21, 22-23, and 24-25. This is a behavior that's common in the dialog, but almost always is without clear indication as to when it happens. Anyway, line 18 in the original Japanese just says "uh-oh" or "oops", but I tried to be crafty and use it to include extra words to not bump into the character limit. Well, this worked fine until coming across lines 24 & 25 which as I found out today are said not ONLY when you try to sell an already-equipped item to the smithy, but those lines also get called when trying to sell from your item box. This means I would entirely have to reorder the structure of all of these lines which might not even be feasible, or try to make a compromise that sacrifices a little bit of character in order to make lines 24-25 be able to fit both when selling to the smithy and from your item box. This didn't get caught in playtesting but luckily it was caught pretty quick after first release, so I'm making this update now to hopefully fix the issue.
I also fixed a couple of smithy lines to be internally consistent with game terms when they weren't before: I'm used to the modern MH "Forge" and "Upgrade" which in 1st gen was "Craft" and "Improve" so I also fixed that mistake. Hopefully not many more of these situations come up
(p.s. also fixed a minor typo)